pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ciept-r"
version = "0.1.0"
description = "Bootstrap research repository for the CIEPT-R project."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
  "torch>=2.10"
]

[project.optional-dependencies]
baseline = [
  "recbole>=1.2.1",
  "scipy>=1.10,<1.14",
  "lmdb>=1.7",
  "tensorboardX>=2.6",
  "geomloss>=0.2.6",
  "setproctitle>=1.3.4",
]
local_models = [
  "transformers>=5.3.0",
  "sentence-transformers>=5.2.3",
  "qwen-vl-utils>=0.0.14",
  "paddlepaddle>=3.3.0",
]
dev = [
  "pytest>=9.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
